修改长度可输入小数

+++ app/src/main/java/com/example/lishuai/modedemo/NewUtils/DeleteScanBean.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 //废弃，用完，布头
 public class DeleteScanBean implements Serializable {
-    private int quantity;//长度
+    private double quantity;//长度
     private int spreadingCount;//次数
     private int type;//次数
     private List<ScanBean> fabrics;
@@ -18,11 +18,11 @@ public class DeleteScanBean implements Serializable {
         this.type = type;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
